+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.11.1 (Upcoming Release)
-- Concrete types of X and y changed to AbstractMatrix{Float64} and AbstractVector{Float64}
 
+- Concrete types of X and y changed to AbstractMatrix{Float64} and AbstractVector{Float64}
+- Change function signatures from ::Function to ::F where {F <: Function}
 
 # v0.11.0